Fix implementation of pointer de-referencing
Since in the JVM all variables of a certain class type are references
(no value semantic), OJG.Deref was simply invoking OJG.load. Given our
delayed code generation approach, we should delay this load when the
referred value is needed, so OJG.Deref should be deleted as a OJG.load
is guaranteed to happen later.