Symbol file format: improve cyclic imports detection
The symbol file, of the current module being compiled, must include the
descriptors of all the imported modules. This enables the detection of
cycles even when types of the involved modules are not re-exported
(re-exporting a type, in fact, triggers the emission of the type's
module descriptor). See the tests for examples.