Fix formal parameter type checking bug
When a case statement is discriminating a case variable among different
types, each case label identifies the type which the case variable will
have for the statements that follow. When the case variable is a formal
parameter of an enclosing procedure, the compiler was erroneously
changing the type of the procedure signature as well. See the added
test for more information.